(From suzukis@file.phys.tohoku.ac.jp)
It seems that current rpm can work on FreeBSD, on OpenBSD,
and possibly on NetBSD (myself have not checked yet), but
find-provides and find-requires scripts are not provided
in rpm tarball. And, taking a glance on the "port"ed version
of rpm by FreeBSD people, they didn't write their own scripts.

Here's find-provides and find-requires for FreeBSD & OpenBSD.
I checked on FreeBSD-2.2.8 and OpenBSD-2.5 (possibly works on 2.6 too)
both a.out executable format. One of the known problem is
that following script cannot check the inter-library dependency.
On both systems, objdump is not bundled in default installation,
and even if you install objdump of GNU binutils,
it cannot check the inter-library dependency.

The recent FreeBSD and NetBSD are in ELF format, so should
be modified... I will send a version for FreeBSD-3.4 sooner
or later, but if anybody is interested in, please write.
I wish these scripts will be included in future version of rpm.

suzuki