here's a patch to make rpm (3.1-0.14) work on current FreeBSD versions (tried 4.0-RELEASE and 5.0-CURRENT).
It adds freebsdelf.prov and freebsdelf.req in autodeps, modifies brp-strip
and brp-strip-comment-note to work with find versions that don't support
-perm +something and strip versions that fail with an error when they
can't strip a file, and modifies the configure check for GNU glob to check
for GLOB_PERIOD additionally.
This is required because FreeBSD's glob has some (but not all) of the GNU
extensions, but doesn't work the way rpm assumes (segfault as soon as
globbing is used in %files); GLOB_PERIOD can be used to tell the BSD
version apart from the GNU one.