Temporarily(!) remove the _partial_ Lua/libsyck support from the build environment to simplify the forthcoming migration to the more powerful RPM_CHECK_LIB-based third-party library linking mechanism.
The current support for libsyck was still not sufficient and complete as
the code even lacks a call to luaopen_syck(), the Lua/* Makefiles do not
contain the necessary glue for building lua/local/lsyck.c, etc. Hence I
think it makes sense to at least temporarily remove the unfinished glue
for now to remove confusion and to simplify the migrations.

Once the more flexible RPM_CHECK_LIB macro is present, we can use it to
easily re-integrate libsyck again into RPM's Lua build environment. But
then certainly together _with_ the necessary code and Makefile glue, of
course.