Another round on my long road of untwisting and simplify the RPM build- and install-environment. This commit addresses the fact that RPM 5 until now used versioned paths in a partly hard-coded and partly unfinished way (mainly just for the lib/rpm/ directory). Instead of hard-coding the versioning in macros and even the C code, move it back to where it really belongs to IMHO: the Autoconf glue. And activate it there only if explicitly requested with option --with-path-versioned.
This way one is again able to build RPM 5 in the way a distribution
vendor finally wishes it (usually without any versioning at all) and the
way a developer might wishes it (usually with full versioning to not
destroy the system RPM installation at all).

                             - - -

Caveat: Automake doesn't support versioned installation paths very good.
Well, actually it doesn't support it at all except for binaries via a
"transform" rule. The only (and even officially documented) way is to
install regularily and then adjust the filesystem files afterwards. This
works just fine, but only if one installs to a *non-system* "prefix" or
at least uses "DESTDIR" to install into an intermediate directory first
and then move the stuff over the system location. So, if you are using
--prefix=/usr please use "make install DESTDIR=/tmp/usr" and _then_
copy over the /tmp/usr files to /usr. For all other "prefix" it usually
doesn't matter.