Ok, let the crowd win: use implicit internal and external library dependencies by default. But at least allow one to explicitly disable it again via --disable-build-intlibdep and --disable-build-extlibdep.
Additionally, remove the GNU libtool *.la files only during
installation if --disable-build-lafiles is used. This is related to the
--disable-build-{int,ext}libdep as for static libraries one can only
"encode" the library interdependencies in the GNU libtool *.la files.

Notice: I personally think this is the wrong direction for RPM 5,
because this way one again receives the possibility to "just link
against librpmxxx", let this magically work and this way especially
circumvent the use of the central pkg-config(1) API (which is IMHO
more consistent as it also provides the path to the includes, etc).
But as I'm alone in my camp, here it is. This is what people wish, not
what IMHO really makes sense. I personally think this is the wrong
direction as it provides _multiple_ ways of linking against the RPM 5
C API, each with its pros and cons and especially each one not 100%
compatible to the others. Unfortunately, my previously implemented "use
the pkg-config(1) API only" has not reached consenus... ;-)