OpenPKG-specific: Use bsdtar(1) for unpacking ZIP format files.
The reason is that OpenPKG uses bsdtar(1) to unpack ZIP format files to
avoid having to bundle the less portable unzip(1) into the bootstrap
package.

Notice: If RPM would use "%__unzip -qq" instead of "%{__unzip} -qq" one
could also use plain macros in OpenPKG to convert the call to unzip(1)
to bsdtar(1). But I don't wanted to change RPM here as all other macros
are also just for the expansion of the command and not for the complete
command argument line.