rpmdb.c (rpmdbInitIterator, rpmdbNextIterator): medium redesign
This fixes infinite loop when RPMDBI_PACKAGES iterator gets damaged header.

The idea is that we should discriminate strictly between mi_set
iterators and DB_NEXT iterators.  That is, when rpmdbInitIterator
is called with non-NULL key, that *must* be mi_set iterator (that is,
mi->mi_set is non-NULL iff key was non-NULL).  Otherwise, the only
sane possibility left is RPMDBI_PACKAGES iterator to traverse headers
(it is not possible to use NULL for other tags, except for the case
when rpmdbGrowIterator is called to populate mi_set after that.)

As long as the distinction between mi_set and DB_NEXT iterators
is reliable, we don't need to store mi_keyp any longer, because
the key will not be used again neither for mi_set nor for DB_NEXT
lookup. mi_keyp and mi_keylen are gone.

mi_key and mi_data should be gone, too.  The do *not* store any
state between iterator calls, which I found misleading.

mi_setx is now used *only* with mi_set.

Special handling of (mi_offset == 0), that is, skipping header
instance #0, has been kept only for DB_NEXT traversal.

"goto top" has been replaced with tail-recursive call.